refactor(artifacts): separate reusable schemas and contracts from app workflows
Apply the #556 top-level family ownership policy to the artifacts family,
mirroring the evaluation relocation (#557). ArtifactMeta and AssetRef move
to archetype.artifacts.components; the processor protocol, table-name and
envelope constants, content/identity digest helpers, and publication
receipts move to archetype.artifacts.contracts; the bundle value contracts
move to archetype.artifacts.bundles. All definitions, defaults, schemas,
and digest outputs are byte-identical.

PreparedArtifactBundleRequest stays in archetype.app.artifacts.bundle_models
because its validation binds to the control catalog's publication-key
derivation (domain archetype.catalog.v1) — publication and storage identity
remain app-owned, and the family never imports app.

Closes #558
