+++ evals/infra/idempotency_worker.py
@@ -19,11 +19,11 @@ from pathlib import Path
 
 from uuid_utils import uuid7
 
-from archetype.app.artifacts.models import ArtifactMeta
 from archetype.app.container import ServiceContainer
 from archetype.app.evaluation.models import EvalReceipt, GraderContract, Outcome
 from archetype.app.gateway.auth.models import ActorCtx
 from archetype.app.storage.catalog import ClaimConflictError, SqliteControlCatalog
+from archetype.artifacts.components import ArtifactMeta
 from archetype.core.component import Component
 from archetype.core.config import RunConfig, StorageConfig, WorldConfig
 from archetype.core.interfaces import StaleWriterError
