+++ evals/suites/capability/indexed_finalization.py
@@ -14,13 +14,7 @@ from pathlib import Path
 from typing import Any
 
 from archetype.app.application.mission_artifacts import MissionArtifactFinalizer
-from archetype.app.artifacts.bundle_models import (
-    ArtifactBundleRequest,
-    ArtifactPublicationStatus,
-    ArtifactPublishReceipt,
-    ArtifactStoreConfig,
-    PreparedArtifactBundleRequest,
-)
+from archetype.app.artifacts.bundle_models import PreparedArtifactBundleRequest
 from archetype.app.artifacts.bundle_service import ArtifactBundleService
 from archetype.app.missions import (
     AttemptClaimAcquireOutcome,
@@ -35,6 +29,12 @@ from archetype.app.missions import (
 )
 from archetype.app.redaction.service import RedactionService
 from archetype.app.storage.catalog import SqliteControlCatalog
+from archetype.artifacts.bundles import (
+    ArtifactBundleRequest,
+    ArtifactPublicationStatus,
+    ArtifactPublishReceipt,
+    ArtifactStoreConfig,
+)
 from archetype.core.config import StorageConfig
 from evals.graders import state_check
 from evals.harness import EvalHarness
