+++ evals/suites/idempotency/durable.py
@@ -17,7 +17,6 @@ from unittest.mock import patch
 
 from uuid_utils import uuid7
 
-from archetype.app.artifacts.models import ArtifactMeta
 from archetype.app.container import ServiceContainer
 from archetype.app.evaluation.models import EvalReceipt, GraderContract, Outcome
 from archetype.app.gateway.auth.models import ActorCtx
@@ -28,6 +27,7 @@ from archetype.app.storage.catalog import (
     WorldRecord,
     claim_scope_key,
 )
+from archetype.artifacts.components import ArtifactMeta
 from archetype.core.component import Component
 from archetype.core.config import RunConfig, StorageConfig, WorldConfig
 from archetype.core.interfaces import StaleWriterError
