+++ src/archetype/__init__.py
@@ -177,36 +177,21 @@ _EXPORTS: dict[str, tuple[str, str]] = {
     "EpisodeResult": ("archetype.app.models", "EpisodeResult"),
     "RolloutConfig": ("archetype.app.models", "RolloutConfig"),
     "RolloutResult": ("archetype.app.models", "RolloutResult"),
-    "ArtifactReceipt": ("archetype.app.artifacts.models", "ArtifactReceipt"),
-    "ArtifactWriteReceipt": ("archetype.app.artifacts.models", "ArtifactWriteReceipt"),
-    "ArtifactProcessor": ("archetype.app.artifacts.models", "ArtifactProcessor"),
-    "ArtifactBundleRequest": (
-        "archetype.app.artifacts.bundle_models",
-        "ArtifactBundleRequest",
-    ),
-    "ArtifactCandidate": ("archetype.app.artifacts.bundle_models", "ArtifactCandidate"),
-    "ArtifactIndexRecord": ("archetype.app.artifacts.bundle_models", "ArtifactIndexRecord"),
-    "ArtifactPublishReceipt": (
-        "archetype.app.artifacts.bundle_models",
-        "ArtifactPublishReceipt",
-    ),
-    "ArtifactReconcileResult": (
-        "archetype.app.artifacts.bundle_models",
-        "ArtifactReconcileResult",
-    ),
-    "ArtifactSourceResolver": (
-        "archetype.app.artifacts.bundle_models",
-        "ArtifactSourceResolver",
-    ),
+    "ArtifactReceipt": ("archetype.artifacts.contracts", "ArtifactReceipt"),
+    "ArtifactWriteReceipt": ("archetype.artifacts.contracts", "ArtifactWriteReceipt"),
+    "ArtifactProcessor": ("archetype.artifacts.contracts", "ArtifactProcessor"),
+    "ArtifactBundleRequest": ("archetype.artifacts.bundles", "ArtifactBundleRequest"),
+    "ArtifactCandidate": ("archetype.artifacts.bundles", "ArtifactCandidate"),
+    "ArtifactIndexRecord": ("archetype.artifacts.bundles", "ArtifactIndexRecord"),
+    "ArtifactPublishReceipt": ("archetype.artifacts.bundles", "ArtifactPublishReceipt"),
+    "ArtifactReconcileResult": ("archetype.artifacts.bundles", "ArtifactReconcileResult"),
+    "ArtifactSourceResolver": ("archetype.artifacts.bundles", "ArtifactSourceResolver"),
     "BoundedArtifactSourceResolver": (
-        "archetype.app.artifacts.bundle_models",
+        "archetype.artifacts.bundles",
         "BoundedArtifactSourceResolver",
     ),
-    "ArtifactStoreConfig": ("archetype.app.artifacts.bundle_models", "ArtifactStoreConfig"),
-    "MaterializedArtifact": (
-        "archetype.app.artifacts.bundle_models",
-        "MaterializedArtifact",
-    ),
+    "ArtifactStoreConfig": ("archetype.artifacts.bundles", "ArtifactStoreConfig"),
+    "MaterializedArtifact": ("archetype.artifacts.bundles", "MaterializedArtifact"),
     # Command models used by supported host adapters
     "Command": ("archetype.app.models", "Command"),
     "CommandType": ("archetype.app.models", "CommandType"),
