+++ src/archetype/app/artifacts/bundle_service.py
@@ -30,7 +30,18 @@ from daft.functions import download, guess_mime_type, upload
 from uuid_utils import uuid7
 
 from archetype import _obs
-from archetype.app.artifacts.bundle_models import (
+from archetype.app.artifacts.bundle_models import PreparedArtifactBundleRequest
+from archetype.app.limits import MAX_ICEBERG_SNAPSHOT_ID
+from archetype.app.redaction.interfaces import iRedactionService
+from archetype.app.redaction.models import RedactionReceipt
+from archetype.app.storage.catalog import (
+    ArtifactPublicationExpiredError,
+    ArtifactPublicationRecord,
+    artifact_publication_key,
+)
+from archetype.app.storage.interfaces import iStorageService
+from archetype.app.world.interfaces import iWorldService
+from archetype.artifacts.bundles import (
     ArtifactBundleRequest,
     ArtifactCandidate,
     ArtifactIndexRecord,
@@ -44,19 +55,8 @@ from archetype.app.artifacts.bundle_models import (
     ArtifactStoreConfig,
     BoundedArtifactSourceResolver,
     MaterializedArtifact,
-    PreparedArtifactBundleRequest,
     _canonical_json,
 )
-from archetype.app.limits import MAX_ICEBERG_SNAPSHOT_ID
-from archetype.app.redaction.interfaces import iRedactionService
-from archetype.app.redaction.models import RedactionReceipt
-from archetype.app.storage.catalog import (
-    ArtifactPublicationExpiredError,
-    ArtifactPublicationRecord,
-    artifact_publication_key,
-)
-from archetype.app.storage.interfaces import iStorageService
-from archetype.app.world.interfaces import iWorldService
 from archetype.core.config import StorageConfig
 
 if TYPE_CHECKING:
