+++ tests/app/test_mission_artifact_finalizer.py
@@ -9,13 +9,7 @@ import pytest
 
 from archetype.app.application import mission_artifacts
 from archetype.app.application.mission_artifacts import MissionArtifactFinalizer
-from archetype.app.artifacts.bundle_models import (
-    ArtifactBundleRequest,
-    ArtifactPublicationStatus,
-    ArtifactPublishReceipt,
-    ArtifactStoreConfig,
-    PreparedArtifactBundleRequest,
-)
+from archetype.app.artifacts.bundle_models import PreparedArtifactBundleRequest
 from archetype.app.artifacts.bundle_service import ArtifactBundleService
 from archetype.app.container import ServiceContainer
 from archetype.app.limits import MAX_ICEBERG_SNAPSHOT_ID
@@ -25,6 +19,12 @@ from archetype.app.missions.models import (
     MissionAttemptRequest,
 )
 from archetype.app.redaction import RedactionService
+from archetype.artifacts.bundles import (
+    ArtifactBundleRequest,
+    ArtifactPublicationStatus,
+    ArtifactPublishReceipt,
+    ArtifactStoreConfig,
+)
 from archetype.core.config import StorageConfig
 from archetype.missions.transitions import (
     FinalizationPhase,
